+++ tests/apis/rules/src/main.c
@@ -4,6 +4,5 @@ extern int test(int a, int b);
 int main(int argc, char** argv)
 {
     printf("1 + 1 = %d\n", test(1, 1));
-    printf("%s\n", TEST);
     return 0;
 }