+++ tests/apis/rules/xmake.lua
@@ -1,6 +1,7 @@
 
 -- define rule: markdown
 rule("markdown")
+    set_extensions(".md", ".markdown")
     on_build(function (target, sourcefile)
         os.cp(sourcefile, path.join(target:targetdir(), path.basename(sourcefile) .. ".html"))
     end)
@@ -44,14 +45,15 @@ rule("stub")
 -- define target
 target("test")
 
-    add_cxflags("-DTEST=\"D:/asdasd/adasd/adsad\"")
-
     -- set kind
     set_kind("binary")
+
+    -- add rules
+    add_rules("markdown")
 
     -- add files
     add_files("src/*.c") 
     add_files("src/man/*.in",   {rule = "man"})
-    add_files("src/index.md",   {rule = "markdown"})
+    add_files("src/index.md")
     add_files("src/test.c.in",  {rule = "c code"})
     add_files("src/empty.stub", {rule = "stub"})